tutorials/agents/requirements.txt
dspy-ai
datasets
ollama
#Add other dependencies like mlflow if you use them
mlflow>=2.20